Fix warnings and unnoticed compile errors.
git-svn-id: https://testsrhoh3aga.myversioncontrol.com/subversion/eahoehaaa2245/trunk@344 e2d0575d-2f82-4250-b8ec-0d5e7f3c5cc0

+++ src/arcemu-world/Player.cpp
@@ -1136,7 +1136,7 @@ void Player::_EventAttack(bool offhand)
 		}
 		setAttackTimer(300, offhand);
 	}
-	else if(!HasInArc(2*M_PI/3, pVictim->GetPositionX(), pVictim->GetPositionY()))
+	else if(!HasInArc(float(2*M_PI/3), pVictim->GetPositionX(), pVictim->GetPositionY()))
 	{
 		// We still have to do this one.
 		if(m_AttackMsgTimer != 2)
